[ARM] Switch to default sched pressure algorithm
Currently the Arm backend selects the alternative sched pressure algorithm.
The issue is that this doesn't take register pressure into account, and so
it causes significant additional spilling on Arm where there are only 14
allocatable registers.  Building SPEC2006 showed significant codesize gains
with the default pressure algorithm, so switch back to that.  PR77308 shows
~800 fewer instructions.

SPECINT2006 is ~0.6% faster on Cortex-A57 together with the other DImode
patches. Overall SPEC codesize is 1.1% smaller.

    gcc/
	* config/arm/arm.c (arm_option_override): Don't override sched
	pressure algorithm.


git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@276960 138bc75d-0d04-0410-961f-82ee72b054a4