[Darwin] Update darwin_binds_local_p.
The use of default_binds_local_p had got out of sync with the varasm
changes, this restores the call to be direct.  In practice, we add some
further tests to determine local binding - but this callback is used for
the initial assessments made by default_encode_section_info().

gcc/ChangeLog:

2019-10-15  Iain Sandoe  <iain@sandoe.co.uk>

	* config/darwin.c (darwin_binds_local_p): Update to call
	default_binds_local_p_3 () directly. amend comments.



git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@277009 138bc75d-0d04-0410-961f-82ee72b054a4