* cp-gimplify.c (cp_gimplify_expr): Use get_initialized_tmp_var.
The comment for get_formal_tmp_var says that it shouldn't be used for
expressions whose value might change between initialization and use, and in
this case we're creating a temporary precisely because the value might
change, so we should use get_initialized_tmp_var instead.

I also noticed that many callers of get_initialized_tmp_var pass NULL for
post_p, so it seems appropriate to make it a default argument.

gcc/
	* gimplify.h (get_initialized_tmp_var): Add default argument to
	post_p.
	* gimplify.c (gimplify_self_mod_expr, gimplify_omp_atomic): Remove
	NULL post_p argument.
	* targhooks (std_gimplify_va_arg_expr): Likewise.

git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@277128 138bc75d-0d04-0410-961f-82ee72b054a4